docs: Simplify example
Since Arroyo provides transform and (soon) produce strategies there should
be no reason to actually implement strategies of your own for simple use
cases like the one in the example. Rewrite the example to leverage
the existing strategies. Note: The produce part will be added separately
as part of https://github.com/getsentry/arroyo/pull/109.
