+++ examples/transform_and_produce/README.md
@@ -10,7 +10,7 @@ The process is fairly simple:
 
 - The consumer consumes messages from the `raw-topic`
 - This message is in the form `{"username": "<username>", "password": "<password>"}`
-- The message is submitted to the `HashPasswordStrategy` which hashes the password string
+- The message is submitted to the `TransofrmStrategy` which hashes the password string
 - The credentials are then submitted to the `ProduceStep` which simply produces the given message to `hash-topic`
 - The `ProduceStep` is also responsible for commiting the original offset back for the consumer
 
@@ -18,9 +18,9 @@ The process is fairly simple:
 
 ---
 
-To run this example ensure each of your shells are inside the examples/transfrom_and_produce directory.  To begin, build the session by running `docker-compose up`.  Once Kafka has entered a running state you may continue onto the next section.
+To run this example ensure each of your shells are inside the examples/transfrom_and_produce directory. To begin, build the session by running `docker-compose up`. Once Kafka has entered a running state you may continue onto the next section.
 
-Note you will see `transform_and_produce-arroyo-1 exited with code 0`.  You can safely ignore it.  This is expected behavior.
+Note you will see `transform_and_produce-arroyo-1 exited with code 0`. You can safely ignore it. This is expected behavior.
 
 **Monitoring your raw-topic consumer.**
 
@@ -32,7 +32,7 @@ docker-compose exec kafka kafka-console-consumer --bootstrap-server kafka:9095 -
 
 **Running your raw-topic producer.**
 
-This process produces the messages that will flow through the system.  We will need to return to this shell to enter commands once we've finished setup.
+This process produces the messages that will flow through the system. We will need to return to this shell to enter commands once we've finished setup.
 
 ```shell
 docker-compose exec kafka kafka-console-producer --bootstrap-server kafka:9095 --topic raw-topic
@@ -40,7 +40,7 @@ docker-compose exec kafka kafka-console-producer --bootstrap-server kafka:9095 -
 
 **Monitoring your hashed-topic consumer.**
 
-This will echo every message produced to the `hashed-topic` topic.  This is your post-transformation output consumer.  You should see the transformations performed by Arroyo here.
+This will echo every message produced to the `hashed-topic` topic. This is your post-transformation output consumer. You should see the transformations performed by Arroyo here.
 
 ```shell
 docker-compose exec kafka kafka-console-consumer --bootstrap-server kafka:9095 --topic hashed-topic
@@ -54,6 +54,6 @@ docker-compose run arroyo python /app/script.py
 
 **Seeing it all come together.**
 
-Produce a message by returning to our producer shell.  Provide the producer a message formatted similarly to this one: `{"username": "user1", "password": "Password1!"}`.
+Produce a message by returning to our producer shell. Provide the producer a message formatted similarly to this one: `{"username": "user1", "password": "Password1!"}`.
 
 Now the message should appear in the `raw-topic` shell, followed by another message in the `hashed-topic` shell. The password field of the message in `hashed-topic` should be the SHA256 hash of the password field of the message we manually produced to `raw-topic`.
